Add sortablity to the Grocery List
Extend ListItems to hold and render a list of items that will slide.
Add sorting function to sort the list after a user has sorted an
element, and send that to the API, so that the whole list can
be updated with the new order in the DB.
When adding a new item, add an order index to it so that it will end up
at the end of the list (like the default behaviour from before), and
send that to the API so that it will stay there if the page is
refreshed.
Add a handle to the ListItem to grab for sorting. It's a bit buggy on
mobile interfaces right now.

Make the sorting handle not display when on a tab other than "All
Items". Sorting only part of the list would be someone of an undefined
function, so we just won't allow it.
Rework the CSS layout so that the destroy button correctly flexes around
the sorting handle.
This CSS rework means we can cleanup a bunch of the margin and absolute
distance hacks to allow for much simpler floating of objects vertically
center and margin distances from each other.
Also fix editing box being a fixed width, now it fits the list size,
while still having a left buffer so that the text lines up with other
lines.